+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.0.7] - 2018-02-27
 ### Fixed
+- Describe missing `--trigger` in help text
 - Build directory owner and attributes for deploy copy mode
 - Do not capture artifacts files after failed step